# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3000
info:
  contact:
    email: support@split-go.com
    name: Split Go API Support
    url: http://www.split-go.com/support
  description: 分帳系統 API 文檔 - 支援群組管理、交易記錄、結算功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Split Go API
  version: "1.0"
paths:
...
  /users/fcm-token:
    post:
      consumes:
      - application/json
      description: 更新用戶的 Firebase Cloud Messaging 推播令牌
      parameters:
      - description: FCM Token 資料
        in: body
        name: request
        required: true
        schema:
          properties:
            fcm_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: FCM Token 更新成功
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤或 FCM Token 為空
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 更新失敗
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新 FCM 推播令牌
      tags:
      - 用戶
  /users/me:
    get:
      description: 獲取當前登入用戶的個人資料
      produces:
      - application/json
      responses:
        "200":
          description: 用戶資料
          schema:
            properties:
              data:
                properties:
                  avatar:
                    type: string
                  email:
                    type: string
                  id:
                    type: integer
                  name:
                    type: string
                  username:
                    type: string
                type: object
              error:
                type: boolean
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "404":
          description: 用戶不存在
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 獲取個人資料
      tags:
      - 用戶
    put:
      consumes:
      - application/json
      description: 更新當前登入用戶的個人資料
      parameters:
      - description: 更新資料
        in: body
        name: request
        required: true
        schema:
          properties:
            avatar:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            properties:
              data:
                properties:
                  avatar:
                    type: string
                  email:
                    type: string
                  id:
                    type: integer
                  name:
                    type: string
                  username:
                    type: string
                type: object
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新個人資料
      tags:
      - 用戶
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 在 header 中加入 "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"